Cleanup & primitive changes
Changed primitive objects to more necessary ones; and small cleanup.

Initially this just to rework the a broken branch issue as well.

+++ #633 - Server/combat/com/rs/game/npc/combat/CombatScript.java
@@ -72,7 +72,7 @@ public abstract class CombatScript {
     }
 
     public static int getRandomMaxHit(NPC npc, int maxHit, int attackStyle, Entity target) {
-	int[] bonuses = npc.getBonuses();
+	short[] bonuses = npc.getBonuses();
 	double att = bonuses == null ? 0 : attackStyle == NPCCombatDefinitions.RANGE ? bonuses[CombatDefinitions.RANGE_ATTACK] : attackStyle == NPCCombatDefinitions.MAGE ? bonuses[CombatDefinitions.MAGIC_ATTACK] : bonuses[CombatDefinitions.STAB_ATTACK];
 	double def;
 	if (target instanceof Player) {
