+++ #633 - Server/combat/com/rs/game/npc/godwars/bandos/GodwarsBandosFaction.java
@@ -18,7 +18,7 @@ public class GodwarsBandosFaction extends NPC {
 
 	public GodwarsBandosFaction(int id, WorldTile tile, int mapAreaNameHash, boolean canBeAttackFromOutOfArea,
 			boolean spawned) {
-		super(id, tile, mapAreaNameHash, canBeAttackFromOutOfArea, spawned);
+		super((short) id, tile, (byte) mapAreaNameHash, canBeAttackFromOutOfArea, spawned);
 	}
 
 	@Override
